+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.0.14 25-06-2021
+
+* added property statusOnIconOffset
+* added property statusOffIconOffset
+
+## 0.0.13
+
+* added property isEnabled
+
 ## 0.0.12
 
 * fixed example layout